requirements.txt
# Основные зависимости для проекта EcoAssist Fox/Wolf Classifier

# PyTorch и глубокое обучение
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
ultralytics>=8.0.0  # Для YOLOv8 классификации (опционально)

# Обработка данных
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0  # Аугментации
pandas>=2.0.0

# Экспорт и оптимизация моделей
onnx>=1.14.0
onnxruntime>=1.15.0
onnxoptimizer>=0.3.0

# Метрики и валидация
scikit-learn>=1.3.0
scipy>=1.11.0

# REST API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.1.0

# Обработка изображений
imageio>=2.31.0
imageio-ffmpeg>=0.4.8  # Для работы с видео

# Утилиты
tqdm>=4.66.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Docker и деплой
gunicorn>=21.0.0
